Implement aux allocator and more
Implemented via overloading operators new/delete.
For highly-compressed allocations that don't change.
KEYD_AUX_POOL overrides max size (can be zero).
Lazy state saving on IPC bind (compressed in aux pool).
Fixed unbind_all, removed push/pop stuff because too lazy.
Refactored kbd_eval to move some stuff outsize.
Got rid of "ephemeral_config".
Fix LED state restoring.

+++ src/keyd.h
@@ -90,7 +90,7 @@ struct ipc_message {
 
 	uint32_t timeout;
 	char data[MAX_IPC_MESSAGE_SIZE];
-	size_t sz;
+	uint64_t sz;
 };
 
 int monitor(int argc, char *argv[]);
